+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## [Unreleased]
 
 - Add support for web downloads and usenet
-- Add download status indicator (Ready / progress percentage)
+- Add support for queued downloads
+- Add download status indicator (Ready / progress percentage / Queued)
 - Add "Refresh All Downloads" action (Cmd+R)
 - Add "Delete Download" action (Ctrl+X)
 - Show all downloads instead of last 10